requirements.txt
alembic==1.0.8
astroid==2.2.5
Babel==2.6.0
blinker==1.4
Click==7.0
colorama==0.4.1
dominate==2.3.5
elasticsearch==7.0.0
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.8
MarkupSafe==1.1.1
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.3.1
pylint-flask==0.6
pylint-plugin-utils==0.5
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.2
typed-ast==1.3.1
urllib3==1.24.2
visitor==0.1.3
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.2.1
